Rename truffle-require to @truffle/require
