+++ packages/require/test/lib/module_with_global_require.js
