+++ packages/require/test/lib/module_with_local_module_require.js
@@ -1,4 +1,4 @@
-// mocha is a dependency of truffle-require, thus it is locally installed.
+// mocha is a dependency of @truffle/require, thus it is locally installed.
 var mocha = require("mocha");
 
 module.exports = function() {
