+++ packages/truffle-core/lib/commands/exec.js
@@ -39,7 +39,7 @@ const command = {
     const Config = require("truffle-config");
     const Contracts = require("@truffle/workflow-compile");
     const ConfigurationError = require("../errors/configurationerror");
-    const Require = require("truffle-require");
+    const Require = require("@truffle/require");
     const { Environment } = require("truffle-environment");
     const path = require("path");
     const OS = require("os");
