+++ packages/truffle-core/package.json
@@ -59,7 +59,7 @@
     "truffle-migrate": "^3.0.31",
     "truffle-provider": "^0.1.15",
     "truffle-provisioner": "^0.1.5",
-    "truffle-require": "^2.0.19",
+    "@truffle/require": "^2.0.19",
     "truffle-resolver": "^5.0.15",
     "@truffle/workflow-compile": "^2.1.3",
     "universal-analytics": "^0.4.17",
