+++ packages/truffle-migrate/migration.js
@@ -1,6 +1,6 @@
 const path = require("path");
 const Deployer = require("truffle-deployer");
-const Require = require("truffle-require");
+const Require = require("@truffle/require");
 const Emittery = require("emittery");
 const { Web3Shim } = require("truffle-interface-adapter");
