+++ packages/truffle-migrate/package.json
@@ -2,7 +2,7 @@
   "name": "truffle-migrate",
   "description": "On-chain migrations management",
   "license": "MIT",
-  "author": "Tim Coulter <tim.coulter@consensys.net>",
+  "author": "Tim Coulter <tim@trufflesuite.com>",
   "homepage": "https://github.com/trufflesuite/truffle-migrate#readme",
   "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/truffle-migrate",
   "bugs": {
@@ -22,7 +22,7 @@
     "@truffle/expect": "^0.0.9",
     "truffle-interface-adapter": "^0.2.4",
     "@truffle/reporters": "^1.0.12",
-    "truffle-require": "^2.0.19",
+    "@truffle/require": "^2.0.19",
     "web3": "1.2.1"
   },
   "devDependencies": {
